Notebook Code For The ProJect
In this Note< Every Step is separated and briefly explained. In this project we use Spatial techniques like Moran I , LISA and Clustering to analyze the accessiblity of public libraries in a particular area or state. This notebook extracts that which areas are underserved or Overserved with in a State.